pyproject.toml
[project]
name = "ampersand"
version = "0.1.1"
description = "Set display background with an image"
readme = {file = "README.org", content-type = "text/org"}
authors = [
    { name = "Brett Viren", email = "brett.viren@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "pillow>=11.3.0",
    "sh>=2.2.2",
    "xlib>=0.21",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project.scripts]
ampersand = "ampersand.__main__:main"
